Fix typing errors

+++ effect.py
@@ -5,10 +5,10 @@ class Effect:
         self.name = name
         self.duration = duration
         self.dps = dps
-    
+
     def is_finished(self) -> bool:
         return self.duration <= 0
-    
+
     def damage(self, target: Entity, delta: int) -> None:
         seconds: float = 0
         if delta < self.duration:
@@ -16,4 +16,4 @@ class Effect:
             seconds = delta / 100
         else:
             seconds = self.duration / 1000
-        target.damage(self.dps * seconds)
+        target.damage(int(self.dps * seconds))
